handle except

+++ client/bob/run_bob.sh
@@ -8,6 +8,7 @@ docker cp ./socket_client_bob.py client_b:/opt/socket/ # copy socket python file
 docker cp ./socket_client_alice.py client_b:/opt/socket/ # copy socket python file into docker
 
 ./../kat_kem $KEM_ALG client_b
+
 mv public_key.txt shared/
 
 
@@ -18,10 +19,15 @@ docker cp client_b:/opt/socket/cipher_text.txt ./shared
 
 #./../kat_kem $KEM_ALG client_b shared/cipher_text.txt secret_key.txt
 ./../kat_kem $KEM_ALG client_b shared/cipher_text.txt secret_key.txt
+if [ $? != 0 ]
+then
+    exit
+fi
 
 #cat decrypted.txt | docker run --rm -it openquantumsafe/curl curl -k https://13.125.30.230:4433/qkey?v=
 #xxd decryted.txt | docker run --rm -it openquantumsafe/curl curl -k https://13.125.30.230:4433/qkey?v=
 #docker run --rm openquantumsafe/curl curl -k https://13.125.30.230:4433/qkey?v=
+
 output=$(python3 encode.py decrypted.txt)
 echo $output
 docker run --rm -it openquantumsafe/curl curl -k https://13.125.30.230:4433/qkey?$1=$output > plaintext.txt
